Primer resultado <3 comparado con ejemplo 1 y 2 con 000,00000 de presicion respectivamente.

+++ src/main/java/LSP/Operadores.java
@@ -12,13 +12,13 @@ public class Operadores {
     private int id;
     private String nombre;
     private String symbol ;
-    private float d;
+    private double d;
     private double r2;
     private double r3;
     private double r4;
     private double r5;
 
-    public Operadores(int id, String nombre, String symbol, float d, double r2, double r3, double r4, double r5) {
+    public Operadores(int id, String nombre, String symbol, double d, double r2, double r3, double r4, double r5) {
         this.id = id;
         this.nombre = nombre;
         this.symbol = symbol;
